Moved app logic to separate file
